requirements.txt
# ============================================
# 核心测试框架
# ============================================
pytest==6.2.4
allure-pytest==2.11.1
allure-python-commons==2.8.34

# ============================================
# 网络请求
# ============================================
requests==2.18.4
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
urllib3==1.22
certifi==2018.1.18
chardet==3.0.4
idna==2.6

# ============================================
# 数据库与缓存
# ============================================
PyMySQL==1.0.2
redis==4.0.2

# ============================================
# 并发处理
# ============================================
gevent==21.8.0
greenlet==1.1.2
zope.event==4.5.0
zope.interface==5.4.0

# ============================================
# 数据解析与序列化
# ============================================
protobuf==4.21.2
blackboxprotobuf==1.0.1
lxml==4.9.0
PyYAML==3.12
jsonpath==0.82
jsonschema==2.6.0
jsonpatch==1.16
jsonpointer==1.10
MarkupSafe==2.1.1
Jinja2==3.1.2

# ============================================
# Git操作
# ============================================
GitPython==3.1.20
gitdb==4.0.7
smmap==4.0.0

# ============================================
# 安全与加密
# ============================================
cryptography>=3.0
PyJWT==1.5.3
pyOpenSSL==17.5.0
pycparser>=2.20

# ============================================
# 工具库
# ============================================
chinesecalendar==1.7.2
future==0.18.2
psutil==5.8.0
websockets==10.3
tornado==6.1
pynsq==0.9.0

# ============================================
# 其他依赖
# ============================================
attrs==21.2.0
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
toml==0.10.2
typing-extensions==3.10.0.0
importlib-metadata==4.6.1
zipp==3.5.0

# ============================================
# 可选/文档
# ============================================
Markdown~=3.4.1
XMind==1.2.0
